+++ tests/fixtures/utils.py
@@ -1,11 +1,12 @@
+from decimal import Decimal
 import json
 from datetime import datetime
-from typing import Type, TypeVar
+from typing import TypeVar
 from uuid import UUID
 
 from sqlalchemy import TIMESTAMP
 from sqlalchemy import UUID as DbUUID
-from sqlalchemy import Boolean, Float, Integer
+from sqlalchemy import Boolean, Numeric, Integer
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.infrastructure.persistence.db.database import Model
@@ -24,8 +25,8 @@ class OrmJsonLoader:
                 data[field_name] = UUID(field_value) if field_value != "None" else None
             if isinstance(column.type, Integer):
                 data[field_name] = int(field_value) if field_value != "None" else None
-            if isinstance(column.type, Float):
-                data[field_name] = float(field_value) if field_value != "None" else None
+            if isinstance(column.type, Numeric):
+                data[field_name] = Decimal(field_value) if field_value != "None" else None
             elif isinstance(column.type, TIMESTAMP):
                 data[field_name] = datetime.fromisoformat(field_value)
             elif isinstance(column.type, Boolean):
